+++ AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.example.infect"
+          package="com.example.funfection"
           android:versionCode="1"
           android:versionName="1.0">
     <uses-sdk android:minSdkVersion="11"/>
@@ -14,7 +14,7 @@
         </activity>
         <activity android:name=".MyVirusActivity">
             <intent-filter>
-                <action android:name="android.inte.action.MAIN"/>
+                <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
